connect to ClientService and VendorService when createEvent

+++ EventService/src/main/java/bezro/model/Event.java
@@ -13,14 +13,14 @@ public class Event {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Integer id;
     private String eventName;
-    private Long clientId;
-    private Long vendorId;
+    private Integer clientId;
+    private Integer vendorId;
     private LocalDateTime eventDateTime;
 
     public Event() {
     }
 
-    public Event(String eventName, Long clientId, Long vendorId, LocalDateTime eventDateTime) {
+    public Event(String eventName, Integer clientId, Integer vendorId, LocalDateTime eventDateTime) {
         this.eventName = eventName;
         this.clientId = clientId;
         this.vendorId = vendorId;
@@ -35,19 +35,19 @@ public class Event {
         this.eventName = eventName;
     }
 
-    public Long getClientId() {
+    public Integer getClientId() {
         return clientId;
     }
 
-    public void setClientId(Long clientId) {
+    public void setClientId(Integer clientId) {
         this.clientId = clientId;
     }
 
-    public Long getVendorId() {
+    public Integer getVendorId() {
         return vendorId;
     }
 
-    public void setVendorId(Long vendorId) {
+    public void setVendorId(Integer vendorId) {
         this.vendorId = vendorId;
     }
